# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_deepakgudla_bookvault_internal_dto.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_deepakgudla_bookvault_internal_dto.OrderItemResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      product:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
      quantity:
        type: integer
    type: object
  github_com_deepakgudla_bookvault_internal_dto.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.OrderItemResponse'
        type: array
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  github_com_deepakgudla_bookvault_internal_dto.ProductImageResponse:
    properties:
      alt_text:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      url:
        type: string
    type: object
  github_com_deepakgudla_bookvault_internal_dto.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductImageResponse'
        type: array
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_deepakgudla_bookvault_internal_utils.PaginatedResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.PaginationMeta'
      success:
        type: boolean
    type: object
  github_com_deepakgudla_bookvault_internal_utils.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_deepakgudla_bookvault_internal_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  internal_server.GraphQLRequest:
    properties:
      operationName:
        type: string
      query:
        example: query { products { edges { node { id name } } } }
        type: string
      variables:
        additionalProperties: true
        type: object
    type: object
  internal_server.GraphQLResponse:
    properties:
      data: {}
      errors: {}
    type: object
host: localhost:1357
info:
  contact:
    email: deepakgudla35@protonmail.com
    name: Deepak Gudla
    url: https://www.linkedin.com/in/deepakgudla
  description: A modern e-commerce API built with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: bookvault-API
  version: "1.0"
paths:
...
  /api/v1/orders/:
    get:
      description: get the list of orders by user
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: orders retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.OrderResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get user's orders
      tags:
      - Orders
    post:
      description: create an order from the current user's cart
      produces:
      - application/json
      responses:
        "201":
          description: successfully created order
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.OrderResponse'
              type: object
        "400":
          description: empty cart or insufficient stock
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create an Order
      tags:
      - Orders
  /api/v1/orders/{id}:
    get:
      description: get the information about the specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: order retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.OrderResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: order not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Orders
...
  /graphql:
    post:
      consumes:
      - application/json
      description: Execute authenticated GraphQL queries and mutations. The GraphQL
        schema defines available operations.
      parameters:
      - description: GraphQL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_server.GraphQLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.GraphQLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.GraphQLResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_server.GraphQLResponse'
      security:
      - BearerAuth: []
      summary: Protected GraphQL endpoint
      tags:
      - GraphQL
  /graphql/public:
    post:
      consumes:
      - application/json
      description: Execute public GraphQL queries. The GraphQL schema defines available
        queries and mutations.
      parameters:
      - description: GraphQL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_server.GraphQLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.GraphQLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.GraphQLResponse'
      summary: Public GraphQL endpoint
      tags:
      - GraphQL
...
securityDefinitions:
  BearerAuth:
    description: Bearer and JWT token (bearer jwt_token)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"